feat: implement tap mapper pattern (#7)
I was running this mapper and I kept hitting the OpenAI API rate limits.
The rate limits are different based on the level of your account but I'm
a "pay as you go" account which is a step up from the free account so if
I'm running into this then free users definitely are too.

When I run a tap-x mapper target-y and the mapper fails its a bad cycle
because I have to start from scratch inevitably erroring out again. I
initially wrote a little backoff logic to handle rate limiting but then
realized this pattern is a solved one in the tap world. This PR is a
hack and I can re-think about where to store my logic better if we like
where its going but it solves an initial need for now.

I'm not sure if this is a terrible pattern but it actually worked pretty
well for me. There was also a similar question in slack
https://meltano.slack.com/archives/C01PKLU5D1R/p1689955015382319
recently where a user needed to tap a source to get a dataset of 200k
user ids that were used as input for their actual tap so they wanted
tap-x tap-y target-snowflake but thats not currently supported, this
might not be a bad solution for them.

@edgarrmondragon @tayloramurphy any thoughts?